d3 force graph with SVG
